\name{bubble_plot}
\alias{bubble_plot}
\title{plot...}
\usage{bubble_plot(x, z, minRadius=0.01, maxRadius=0.05, color="#F75252", ...)}
\description{plot}
\arguments{\item{x}{a SpatialPointsDataFrame}
\item{z}{a vector to be mapped to size}
\item{minRadius}{smallest point size}
\item{maxRadius}{largest point size}
\item{color}{the color of the points}
\item{...}{additional parameters for symbol}}
