\docType{package}
\name{DeducerSpatial-package}
\alias{DeducerSpatial}
\alias{DeducerSpatial-package}
\title{Deducer for spatial data analysis.}
\description{
  Deducer for spatial data analysis.
}
\details{
  \tabular{ll}{ Package: \tab DeducerSpatial\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2011-04-04\cr
  License: \tab LGPL-2\cr LazyLoad: \tab yes\cr }

  A Deducer plug-in for spatial data analysis.
}
\author{
  Ian Fellows \email{ian.fellows@stat.ucla.edu}
}
\references{
  \url{http://www.deducer.org}
}
\keyword{package}

