\name{osmtile}
\alias{osmtile}
\title{an open street map tile.}
\usage{osmtile(x, y, zoom, type="osm")}
\description{an open street map tile. tpe can be "osm" or "bing"}
\value{a tile}
\arguments{\item{x}{location in osm native coordinates}
\item{y}{location in osm native coordinates}
\item{zoom}{zoom level}
\item{type}{osm for mapnik open street map, or 'bing' for bing aerial}}
