\name{plot.OpenStreetMap}
\alias{plot.OpenStreetMap}
\title{the map in mercator coordinates.}
\usage{plot.OpenStreetMap(x, y, add=FALSE, ...)}
\description{the map in mercator coordinates. see osm().}
\arguments{\item{x}{the OpenStreetMap}
\item{y}{ignored}
\item{add}{add to current plot}
\item{...}{additional parameters to be passed to plot}}
