\name{plot.osmtile}
\alias{plot.osmtile}
\title{tile to plot...}
\usage{
\method{plot}{osmtile}(x, y=NULL, add=TRUE, raster=FALSE, ...)
}
\description{tile to plot}
\arguments{\item{x}{the tile}
\item{y}{ignored}
\item{add}{add to current plot}
\item{raster}{plot as raster image}
\item{...}{additional parameters to image}

}
