\name{project_mercator}
\alias{project_mercator}
\title{long lat values to the open street map mercator projection...}
\usage{project_mercator(lat, long, drop=TRUE)}
\description{long lat values to the open street map mercator projection}
\arguments{\item{lat}{a vector of latitudes}
\item{long}{a vector of longitudes}
\item{drop}{drop to lowest dimension}}
