\name{spatialBubblePlot}
\alias{spatialBubblePlot}
\title{bubble plot}
\usage{
spatialBubblePlot(x,z,minRadius=.01,
		maxRadius=.05,color="#F75252", ...)
}
\arguments{
  \item{x}{a SpatialPointsDataFrame}

  \item{z}{The name of the variable in x or a vector to be mapped to size}

  \item{minRadius}{smallest point size}

  \item{maxRadius}{largest point size}

  \item{color}{the color of the points}

  \item{...}{additional parameters for symbol}
}
\description{
  bubble plot
}

