\name{spatialColoredPoints}
\alias{spatialColoredPoints}
\title{plot spatial points with colors}
\usage{
spatialColoredPoints(x, color_var, pch=1, palette , legend.loc="bottomleft",
		legend.title,...)
}
\arguments{
  \item{x}{a SpatialPointsDataFrame}

  \item{color_var}{the name of the variable in x, or a vector}

  \item{pch}{plotting symbol}
  
  \item{palette}{A color scale (See the scales package)}

  \item{legend.loc}{the location of the legend}

  \item{legend.title}{title}

  \item{...}{additional parameters for plot}
}
\description{
  plot spatial points with colors
}

