\name{colored_points}
\alias{colored_points}
\title{plot spatial points with colors}
\usage{
  colored_points(x, color_var, pch = 1,
    legend.loc = "bottomleft", legend.title = NULL, ...)
}
\arguments{
  \item{x}{a SpatialPointsDataFrame}

  \item{color_var}{a vector}

  \item{pch}{plotting symbol}

  \item{legend.loc}{the location of the legend}

  \item{legend.title}{title}

  \item{...}{additional parameters for plot}
}
\description{
  plot spatial points with colors
}

